+++ Rdmp.Core/Databases/CatalogueDatabase/up/091_AddTemplateFlagToCIC.sql
@@ -1,4 +1,4 @@
-﻿--Version: 8.4.3
+﻿--Version: 9.0.4
 --Description: Add new metadata fields for catalogues
 if not exists (select 1 from sys.columns where name = 'IsTemplate' and OBJECT_NAME(object_id) = 'CohortIdentificationConfiguration')
 BEGIN